Tested in dev and fixed issues
