+++ .env.example
@@ -12,8 +12,6 @@ DEV_NAME="SECRET"
 DEV_ID="SECRET"
 DEV_EMAIL="SECRET"
 
-
-
 MAIL_HOST="smtp.ethereal.email"
 MAIL_PORT=587
 MAIL_USER="Your Ethereal email"
